Move service provisioning of process manager to event context
