+++ app/Place/PlaceEditingServiceProvider.php
@@ -3,7 +3,6 @@
 namespace CultuurNet\UDB3\Silex\Place;
 
 use Broadway\UuidGenerator\Rfc4122\Version4Generator;
-use CultuurNet\UDB3\Event\LocationMarkedAsDuplicateProcessManager;
 use CultuurNet\UDB3\Place\Commands\PlaceCommandFactory;
 use CultuurNet\UDB3\Place\DefaultPlaceEditingService;
 use CultuurNet\UDB3\Place\PlaceOrganizerRelationService;
@@ -35,15 +34,6 @@ class PlaceEditingServiceProvider implements ServiceProviderInterface
                 return new PlaceOrganizerRelationService(
                     $app['place_editing_service'],
                     $app['place_relations_repository']
-                );
-            }
-        );
-
-        $app[LocationMarkedAsDuplicateProcessManager::class] = $app->share(
-            function ($app) {
-                return new LocationMarkedAsDuplicateProcessManager(
-                    $app['event_relations_repository'],
-                    $app['event_command_bus']
                 );
             }
         );
